Update FASTFarmCaseCreation.py
Adding functions to create the fstf file from Mann Box. creating functions to create and execute .bat files for TurbSim. Adding floating flag as well as Mooring files and rotating the mooring system with the wind direction. Adding the option to have a different Hydrodyn file for each turbine which is needed for floating wind turbines. There are also some other minor changes.